Refactor provider settings form list into a component
