+++ src/routes/console/project-[project]/messaging/providers/popoverContent.svelte
