+++ src/routes/console/project-[project]/messaging/providers/store.ts
@@ -67,7 +67,7 @@ export const providers: ProvidersMap = {
                         popover: [
                             '<b>How to get the <a class="link" href="https://firebase.google.com" target="_blank" rel="noopener noreferrer">FCM</a> service account JSON?</b>',
                             'Head to <b>Project settings -> Service accounts -> Generate new private key.</b>',
-                            'Generating the new key will result in the download of a JSON file.',
+                            'Generating the new key will result in the download of a JSON file.'
                         ],
                         popoverImage: {
                             src: {
@@ -533,3 +533,124 @@ export const providers: ProvidersMap = {
         }
     }
 };
+
+type ProviderParams = {
+    providerId: string;
+    name: string;
+    enabled: boolean;
+};
+
+/**
+ * SMS providers
+ */
+
+export type TwilioProviderParams = ProviderParams & {
+    accountSid: string;
+    authToken: string;
+    from: string;
+};
+
+export type Msg91ProviderParams = ProviderParams & {
+    from: string;
+    senderId: string;
+    authKey: string;
+};
+
+export type TelesignProviderParams = ProviderParams & {
+    from: string;
+    username: string;
+    password: string;
+};
+
+export type TextmagicProviderParams = ProviderParams & {
+    from: string;
+    username: string;
+    apiKey: string;
+};
+
+export type VonageProviderParams = ProviderParams & {
+    from: string;
+    apiKey: string;
+    apiSecret: string;
+};
+
+/**
+ * Email providers
+ */
+
+export type MailgunProviderParams = ProviderParams & {
+    fromEmail: string;
+    fromName: string;
+    replyToEmail: string;
+    replyToName: string;
+    isEuRegion: boolean;
+    apiKey: string;
+    domain: string;
+};
+
+export type SendgridProviderParams = ProviderParams & {
+    fromEmail: string;
+    fromName: string;
+    replyToEmail: string;
+    replyToName: string;
+    apiKey: string;
+};
+
+export type SMTPProviderParams = ProviderParams & {
+    fromEmail: string;
+    fromName: string;
+    replyToEmail: string;
+    replyToName: string;
+    host: string;
+    port: number;
+    username: string;
+    password: string;
+    encryption: SMTPEncryption;
+    autoTLS: boolean;
+    mailer: string;
+};
+
+/**
+ * Push providers
+ */
+
+export type FCMProviderParams = ProviderParams & {
+    serviceAccountJSON: string;
+};
+
+export type APNSProviderParams = ProviderParams & {
+    authKey: string;
+    authKeyId: string;
+    teamId: string;
+    bundleId: string;
+};
+
+export type MQTTProviderParams = ProviderParams & {
+    serverKey: string;
+};
+
+export const providerType = writable<MessagingProviderType>(null);
+export const provider = writable<Providers>(null);
+export const providerParams = writable<{
+    twilio: Partial<TwilioProviderParams>;
+    msg91: Partial<Msg91ProviderParams>;
+    telesign: Partial<TelesignProviderParams>;
+    textmagic: Partial<TextmagicProviderParams>;
+    vonage: Partial<VonageProviderParams>;
+    mailgun: Partial<MailgunProviderParams>;
+    sendgrid: Partial<SendgridProviderParams>;
+    smtp: Partial<SMTPProviderParams>;
+    fcm: Partial<FCMProviderParams>;
+    apns: Partial<APNSProviderParams>;
+}>({
+    twilio: null,
+    msg91: null,
+    telesign: null,
+    textmagic: null,
+    vonage: null,
+    mailgun: null,
+    sendgrid: null,
+    smtp: null,
+    fcm: null,
+    apns: null
+});
